+++ src/lib.rs
@@ -88,10 +88,8 @@
 //!   * plugin presets: `GameOfLife3dPlugin`, `ImmigrationGame3dPlugin`,
 //!     `RainbowGame3dPlugin`, `WireWorld3dPlugin`, `CyclicAutomaton3dPlugin`
 //! * `auto-coloring` (Example or debug purpose):
-//!   * Enables `CellStateMaterials` resource to contain material handles
-//!   * The `CellState` type now requires to build a `CellStateMaterials`
-//!   * All `CellState` components with materials will be colored according to
-//!     their type.
+//!   * The `CellState` trait now requires a `color` method
+//! * `bevy_reflect` (enabled by default): Enable support for reflection for common types
 //!
 //! ## Disclaimer
 //!